Android 16.0.0 release 4

* tag 'android-16.0.0_r4': (391 commits)
  Fix deadlock issue in Pre-reboot Dexopt.
  Revert "Add extra DCHECKs and static_asserts"
  Fast compiler: bail on catch handlers happening before try blocks.
  Add extra DCHECKs and static_asserts
  Madvise recyclable pages during compaction when using MOVE ioctl
  Remove redundant timestamp fetch when recording exit events
  Revert "Tradefed: Generate more descriptive test names"
  Follow-up comments for weak `const-string`.
  Fast compiler: implement spilling.
  x86/x86_64/arm/arm64 nterp: Smaller `op_const_object`.
  Do not strongly intern strings for JIT root table.
  ART: Fix `JitCodeCache::RemoveUnmarkedCode()`.
  Improve the efficiency of dex2oat log redirection.
  Remove unused variable from test 051-
  Add Flags for Safter Native DCL
  ART: Make const-string interns weak.
  Optimizing: Minor refactoring in `RegisterAllocator`.
  Increase size of BitTableBase::column_offset_ to uint32
  Add logs when hit unreachable in FindNthLiveWordOffset
  Tradefed: Generate more descriptive test names
  ...

+++ libartbase/base/bit_table.h
@@ -54,7 +54,7 @@ class BitTableBase {
     column_offset_[0] = 0;
     for (uint32_t i = 0; i < kNumColumns; i++) {
       size_t column_end = column_offset_[i] + header[i + 1];
-      column_offset_[i + 1] = dchecked_integral_cast<uint16_t>(column_end);
+      column_offset_[i + 1] = dchecked_integral_cast<uint32_t>(column_end);
     }
 
     // Record the region which contains the table data and skip past it.
@@ -98,7 +98,7 @@ class BitTableBase {
  protected:
   BitMemoryRegion table_data_;
   uint32_t num_rows_ = 0;
-  uint16_t column_offset_[kNumColumns + 1] = {};
+  uint32_t column_offset_[kNumColumns + 1] = {};
 };
 
 // Helper class which can be used to create BitTable accessors with named getters.
